fastapi-playlists/requirements.txt
fastapi
uvicorn
mysql-connector-python
requests

